+++ libs/tokens/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [3.13.0](https://github.com/cowprotocol/cowswap/compare/tokens-v3.12.2...tokens-v3.13.0) (2026-09-23)
+
+### Minor Changes
+
+- feat: show twaps in explorer (#8129)
+
+- feat(solana): apply user slippage from settings (#8164)
+
+### Patch Changes
+
+- fix(rwa): update ondo tokens (#8182)
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.17.0
+  - @cowprotocol/ui@3.14.0
+  - @cowprotocol/common-const@3.13.0
+  - @cowprotocol/common-hooks@3.13.0
+  - @cowprotocol/core@3.10.0
+  - @cowprotocol/currency@1.6.0
+  - @cowprotocol/types@4.11.0
+
 ## [3.12.2](https://github.com/cowprotocol/cowswap/compare/tokens-v3.12.1...tokens-v3.12.2) (2026-09-15)
 
 ### Patch Changes